Add a color picker component to the form
A new ColorPicker blade component was created to implement a color picker into the form. The forms.components.color-picker view was added to its associate PHP class, with additional functionalities- setting and getting width. This change was supplemented with modification in 'app.js', adding import statement for the newly introduced color picker.

Further, the ColorPicker was inserted into the CategoryResource as part of the form where color selection is necessary. Category resource views for 'edit' and 'create' were also uncommented.

'@jaames/iro' dependency in 'package.json' and 'package-lock.json' was introduced to achieve this functionality. This will give users the ability to choose colors through a visual tool, improving usability and user experience.